+++ lib/widgets/index.dart
@@ -13,3 +13,4 @@ export 't_number_field.dart';
 export 't_image.dart';
 export 't_scrollable_column.dart';
 export 't_search_field.dart';
+export 'widget_types.dart';