+++ lib/widgets/t_image.dart
@@ -2,15 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:t_widgets/t_widgets.dart';
 
 class TImage extends StatelessWidget {
-  String source;
-  String? defaultAssetsPath;
-  BoxFit fit;
-  double? width;
-  double? height;
-  double? size;
-  double borderRadius;
-  Widget? loadingProgressWidget;
-  TImage({
+  final String source;
+  final String? defaultAssetsPath;
+  final BoxFit fit;
+  final double? width;
+  final double? height;
+  final double? size;
+  final double borderRadius;
+  final FilterQuality filterQuality;
+  final LoadingBuilderCallback? loadingBuilder;
+  final FrameBuilderCallback? frameBuilder;
+  final ErrorBuilderCallback? errorBuilder;
+  const TImage({
     super.key,
     required this.source,
     this.defaultAssetsPath,
@@ -19,7 +22,10 @@ class TImage extends StatelessWidget {
     this.height,
     this.borderRadius = 5,
     this.size,
-    this.loadingProgressWidget,
+    this.errorBuilder,
+    this.frameBuilder,
+    this.loadingBuilder,
+    this.filterQuality = FilterQuality.medium,
   });
 
   @override
@@ -33,7 +39,10 @@ class TImage extends StatelessWidget {
         height: height,
         size: size,
         width: width,
-        loadingProgressWidget: loadingProgressWidget,
+        filterQuality: filterQuality,
+        errorBuilder: errorBuilder,
+        frameBuilder: frameBuilder,
+        loadingBuilder: loadingBuilder,
       );
     }
     return TImageFile(
@@ -44,6 +53,9 @@ class TImage extends StatelessWidget {
       height: height,
       size: size,
       width: width,
+      filterQuality: filterQuality,
+      errorBuilder: errorBuilder,
+      frameBuilder: frameBuilder,
     );
   }
 }